+++ src/main/kotlin/io/unthrottled/doki/ui/UIDefaults.kt
@@ -1,5 +1,0 @@
-package io.unthrottled.doki.ui
-
-import javax.swing.plaf.ColorUIResource
-
-val DEFAULT_COLOR = ColorUIResource(ColorUIResource(0x2D79E9))