+++ changelog/RELEASE-NOTES.md
@@ -1,2 +1,39 @@
-- Fixed small 2020.3 Look and Feel issues. [See issue for more details](https://github.com/doki-theme/doki-theme-jetbrains/issues/292)
-- Fixed issue when attempting to promote the [Waifu Motivator Plugin](https://plugins.jetbrains.com/plugin/13381-waifu-motivator). 
+## 5 New Themes!
+
+Love Live! series:
+
+- Sonoda Umi (Dark)
+
+From the OreGairu series:
+
+- Yukinoshita Yukino (Dark)
+
+Addition to Re:Zero series:
+
+- Echidna (Dark)
+
+From the Steins Gate series:
+
+- Makise Kurisu (Dark)
+
+Addition to the Sword Art Online series:
+
+- Yuuki Asuna (Dark)
+
+
+<img src='http://doki.assets.unthrottled.io/misc/v12_girls_smol.png' alt='v12 Girls' />
+
+## Other Stuff
+
+### Updates
+
+- Konata's theme is now a bit darker to aid in usability
+
+### Miscellaneous
+
+- Update Rin's syntax coloring just a bit.
+- Migrated editor scheme color overrides to all themes.
+- "Last Name First Name"'d Misato.
+- Current line number is now `infoForeground` colored for dark themes.
+- Added search field error background color.
+- Removed obsolete File Colors action.